bentoML_summarization/requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentencepiece
bentoml>=1.0.0
prometheus-client
pillow
scikit-learn
tf2onnx
onnxruntime
matplotlib